@@ -1,7 +1,5 @@
 module github.com/boggydigital/pathways
 
-go 1.21
-
-require golang.org/x/exp v0.0.0-20240112132812-db7319d0e0e3
+go 1.23.6
 
 require github.com/boggydigital/wits v0.2.3
